Update params in calls to LD/STREXW to be uint32_t

+++ targets/TARGET_Maxim/TARGET_MAX32630/mxc/mxc_lock.h
@@ -80,12 +80,12 @@ __STATIC_INLINE int mxc_get_lock(uint32_t *lock, uint32_t value)
     do {
 
         // Return if the lock is taken by a different thread
-        if(__LDREXW((volatile unsigned long *)lock) != 0) {
+        if(__LDREXW((volatile uint32_t *)lock) != 0) {
             return E_BUSY;
         }
 
         // Attempt to take the lock
-    } while(__STREXW(value, (volatile unsigned long *)lock) != 0);
+    } while(__STREXW(value, (volatile uint32_t *)lock) != 0);
 
     // Do not start any other memory access until memory barrier is complete
     __DMB();
